Move common test code into the testing module.
